Refactor: Move ApplicationIdentityMigrationService and its repositories from application-service to core
