+++ pre-registration/pre-registration-core/src/main/java/io/mosip/preregistration/core/common/repository/DocumentRepository.java
@@ -2,7 +2,7 @@
  * 
  * Copyright
  */
-package io.mosip.preregistration.application.repository;
+package io.mosip.preregistration.core.common.repository;
 
 import java.util.List;
